Move translate method into module
