+++ lib/komponent/komponent_helper.rb
@@ -18,6 +18,7 @@ module KomponentHelper
     lookup_context.prefixes = ["components/#{component}"]
 
     context.class_eval { prepend component_module }
+    context.class_eval { prepend Komponent::Translation }
 
     capture_block = proc { capture(&block) } if block
 
@@ -72,22 +73,4 @@ module KomponentHelper
 
     rendered_partial
   end
-
-  def translate(key, options = {})
-    virtual_path = @virtual_path
-
-    is_component = key.to_s.first == "." and
-      virtual_path =~ /$components/
-
-    if is_component
-      path = virtual_path.match(/^components\/.+\/_(.+)/)[1]
-      defaults = [:"#{path}#{key}"]
-      defaults << options[:default] if options[:default]
-      options[:default] = defaults.flatten
-      key = "#{path}.#{key}"
-    end
-
-    super(key, options)
-  end
-  alias :t :translate
 end
